Create espada_workout_timer.py
initial commit of code so far.
timer display not working for any labels, process execution has to be divided among functions so the display can update between function calls. 
DOES NOT WORK LIKE HTML because TK inter is a different structure that isn't nice about displaying as it's processing info.